requirements.txt
asyncio
pymongo
python-dotenv
rapidfuzz